+++ previous-versions/banking/pkg/models/shared/transactioncategory.go
@@ -12,9 +12,10 @@ type TransactionCategory struct {
 	// The name of the bank transaction category.
 	Name string `json:"name"`
 	// The unique identifier of the parent bank transaction category.
-	ParentID           *string                    `json:"parentId,omitempty"`
-	SourceModifiedDate *string                    `json:"sourceModifiedDate,omitempty"`
-	Status             *TransactionCategoryStatus `json:"status,omitempty"`
+	ParentID           *string `json:"parentId,omitempty"`
+	SourceModifiedDate *string `json:"sourceModifiedDate,omitempty"`
+	// The status of the transaction category.
+	Status *TransactionCategoryStatus `json:"status,omitempty"`
 }
 
 func (o *TransactionCategory) GetHasChildren() *bool {